add ability to stop automatic overwrite

+++ transformer/io.py
@@ -4,7 +4,10 @@ from typing import Any
 
 class Writer():
     """Writer streams data to a destination file."""
-    def __init__(self, dest: str, fs: AbstractFileSystem) -> None:
+    def __init__(self, dest: str, fs: AbstractFileSystem, ow: bool) -> None:
+        if fs.exists(dest) and not ow:
+            raise FileExistsError(
+                "File {dest} exists and overwrite is not set")
         self.fout = fs.open(dest, 'wb')
 
     def __call__(self, data: Any) -> bool:
